Add AIModule, Test2.2:bom

+++ AIClient/Source/ReinforcementLearning.cpp
@@ -122,7 +122,7 @@ void ReinforcementLearning::stateAction(){
 	sim->stateAction();
 }
 
-int ReinforcementLearning::reward(){
+double ReinforcementLearning::reward(){
 	return sim->reward();
 }
